Tried to implement strsplit to use heap memory, will fix soon.

+++ utils.h
@@ -1,5 +1,5 @@
 void error(char *msg);
 void gethost(char *var);
-int strsplit(char *str, char delim, char **strings, int len);
+int strsplit(char *str, char delim, char **strings, size_t len);
 void runCmd(char **args);
